Struggling with Kivy
Kivy violates the principal of least surprise frequently.  This has led to a lot of frustration on my part as I attempt to get things to look vaguely like I would like.

These resources were the most useful.

Some good fundamentals:
   http://robertour.com/2013/07/19/10-things-you-should-know-about-the-kivy-canvas/

Good explanation of the different types of layouts:
   https://kivy.org/doc/stable/guide/widgets.html#organize-with-layouts

In the end I couldn't get any of the layouts to look the way that I wanted and I've just reverted to using absolute positioning along with parent position for my 'topmost' display elements.  Perhaps this is the right way to go about this anyway.